refactor: optimizing and separating functions by responsibility
